fix: Complete compilation fixes for all examples, binaries, and demos
Key achievements:
- Fixed all import path errors across examples and binaries
- Enhanced lib.rs exports with KTable, KafkaAdminClient, ProducerBuilder
- Updated kafka/mod.rs with proper re-exports for examples
- Fixed schema registry imports in test binaries
- Fixed file data source config imports in demos
- Resolved clippy warnings for unused variables and patterns
- Added spec-driven development templates

All examples, binaries, sources, and demos now compile successfully.
GitHub Actions CI/CD will now pass compilation checks.

+++ src/ferris/serialization/protobuf_codec.rs
@@ -327,7 +327,7 @@ impl ProtobufCodec {
             }
             FieldValue::Decimal(decimal) => {
                 // Convert rust_decimal::Decimal to our DecimalMessage
-                let scale = decimal.scale() as u32;
+                let scale = decimal.scale();
                 let mantissa = decimal.mantissa();
                 // Convert i128 to i64, handling potential overflow
                 let units = if mantissa > i64::MAX as i128 || mantissa < i64::MIN as i128 {
